t1/cache_casero/app/server/requirements.txt
grpcio==1.57.0
grpcio-tools==1.57.0
uhashring==2.3
numpy>=1.21.2

